instaclone/social/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers
from social.models import Post, Comment


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["id", "username", "email"]


class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = "__all__"


class PostSerializer(serializers.ModelSerializer):
    comments = CommentSerializer(many=True, read_only=True)
    user = UserSerializer(read_only=True)

    class Meta:
        model = Post
        fields = ["image", "caption", "created_at", "likes", "user", "comments"]
        read_only_fields = ["user", "created_at"]
