instaclone/requirements.txt
djangorestframework
pillow
